Refactor more source conversion
